+++ CHANGELOG.rst
@@ -61,7 +61,7 @@ Changed
     - Update format strings (#2197)
     - Refactor iRODS path handling (#2218)
 - **Irodsadmin**
-    - Refactor ``irodsorphans`` command (#2246)
+    - Refactor ``irodsorphans`` command (#2207, #2246)
     - Refactor ``irodsorphans`` tests (#2279)
 - **Irodsbackend**
     - Update app for ``public_access`` support (#2202)
@@ -98,6 +98,7 @@ Fixed
     - ``metadata_advanced.rst`` references broken in docs (#2274)
 - **Irodsadmin**
     - Redundant queries for category iRODS paths in ``irodsorphans`` (#2280)
+    - Nested paths set by assay plugin reported as orphaned (#2207)
 - **Samplesheets**
     - Details card iRODS stats badge fails with guest role (#2243, #2271)
     - Sheet Operations dropdown disabled for guest user (#2244)